# docs/swagger.yaml
definitions:
...
  main.SProviderRegistration:
    properties:
      Existed:
        example: false
        type: boolean
      Password:
        example: huefw3fw3
        type: string
      Success:
        example: true
        type: boolean
    type: object
info:
  contact: {}
  description: ControlFloor Server API
  title: ControlFloor API
  version: "1.0"
paths:
...
  /provider/login:
    post:
      description: Provider - Login
      parameters:
      - description: Username
        in: query
        name: user
        required: true
        type: string
      - description: Password
        in: query
        name: pass
        required: true
        type: string
...
  /provider/register:
    post:
      description: Provider - Register
      parameters:
      - description: Registration password
        in: formData
        name: regPass
        required: true
        type: string
      - description: Provider username
        in: formData
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.SProviderRegistration'
...
swagger: "2.0"